Implemented ability to view more details about downloaded files
User can set long click on a downloaded file in the listview to view details about the file which is displayed in a FrameLayout which overlays the ListView when open. From here the user can see the date created, name of file, file size aswell as the abiltity to delete the downloaded file.